+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -137,7 +137,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-15T10:38:53.975Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-15T13:29:06.223Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")