+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KV5DTNB4RBGMQ4DCENF8ZJST"
+        "01KV5QJJKX0CMHZ4WGJHFSBM6F"
     )
     r"""The unique identifier of the data source"""