+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KV5DTNB8FM7AMHW39HWPX50W"
+        "01KV5QJJM1WAKN8ZZ5BWZ50Y9N"
     )
     r"""The unique identifier of the data source"""