+++ packages/orq-rc/RELEASES.md
@@ -8719,3 +8719,13 @@ Based on:
 - [python v4.11.0-rc.21] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.21] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.21 - packages/orq-rc
+
+## 2026-06-15 13:31:57
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.778.0 (2.904.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.22] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.22] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.22 - packages/orq-rc